saracatunga

+++ MissTortasRework/MissTortas.Data/Entity/Products/Product.cs
@@ -6,7 +6,7 @@ namespace MissTortas.Data.Entity.Products
         public string Name { get; set; } = string.Empty;
         public required ProductDetail ProductDetail { get; set; }
         public required ProductCategory ProductCategory { get; set; }
-        public float Quantity { get; set; }
+        public double Quantity { get; set; }
         public string Unit { get; set; } = string.Empty;
     }
 }
