+++ MissTortasRework/MissTortas.Data/Entity/Products/SaleProduct.cs
@@ -5,8 +5,8 @@ namespace MissTortas.Data.Entity.Products
         public long Id { get; set; }
         public required Product StockProduct { get; set; }
         public string SaleDescription { get; set; } = string.Empty;
-        public float SalePrice { get; set; }
-        public long SaleQuantity { get; set; }
+        public double SalePrice { get; set; }
+        public double SaleQuantity { get; set; }
         public string SaleImagePath { get; set; } = string.Empty;
     }
 }
