+++ MissTortasRework/MissTortas.Services/DTO/Products/SaleProductCreateDTO.cs
@@ -3,7 +3,7 @@
     public class SaleProductCreateDTO
     {
         public float SalePrice { get; set; }
-        public long SaleQuantity { get; set; }
+        public float SaleQuantity { get; set; }
         public long ProductId { get; set; }
         public string SaleDescription { get; set; } = string.Empty;
         public string SaleImagePath { get; set; } = string.Empty;
