+++ MissTortasRework/MissTortas.Services/DTO/Products/SaleProductDTO.cs
@@ -5,7 +5,7 @@ namespace MissTortas.Services.DTO.Products
     public class SaleProductDTO
     {
         public long Id { get; set; }
-        public long Quantity { get; set; }
+        public double Quantity { get; set; }
         public double Price { get; set; }
         public string Description { get; set; } = string.Empty;
     }
